+++ R/sample.R
@@ -1,48 +1,74 @@
 # Function -------------------------------------------------------------------
 
-.sample <- function(x, size, replace = FALSE, prob = NULL) {
+.sample_h <- function(x, size, replace = FALSE, prob = NULL, index = FALSE) {
   validObject(x)
 
   if (missing(size)) size <- nrow(x) # normal default does not work!
 
   rows_i <-
     sample.int(nrow(x@data), size = size, replace = replace, prob = prob)
-  x[rows_i]
+
+  if (isTRUE(index)) {
+    rows_i
+  } else {
+    x[rows_i]
+  }
 }
 
-#' Random samples and permutations
+#' Sample or permute rows of `hyperSpec`, `data.frame`, or `matrix`
 #'
-#' [sample()] takes a specified size sample of rows from the object `x` with
-#' or without replacement.
+#' [hyperSprc::sample()] takes a random sample (drawn with or  without
+#' replacement) of rows from the object `x`.
 #'
+#' @name sample
 #' @rdname sample
+#' @aliases sample.hyperSpec
+#'          sample,hyperSpec-method
+#'
 #' @docType methods
 #'
-#' @param x For [sample()]: `hyperSpec` object, `data.frame` or `matrix` to
-#'          sample from.
-#'          For [isample()]: `hyperSpec` object.
+#' @param x `hyperSpec` object, `data.frame` or `matrix` to sample rows from.
+#'
 #' @param size (integer): Positive integer giving the number of spectra (rows)
-#'        to choose. Id missing, defaults to number of rows in the object:
+#'        to choose. If missing, defaults to the number of rows in the object:
 #'        `size = nrow(x)`.
+#'
 #' @param replace (logical): Should sampling be with replacement?
+#'
 #' @param prob (numeric): A vector of probability weights for obtaining the
 #'        elements of the vector being sampled.
 #'
-#' @return
-#' [sample()] returns a `hyperSpec` object, `data.frame` or `matrix` with `size`
-#' rows.
+#' @param index (logical): If `FALSE`, object of `class(x)` is returned,
+#'        if `TRUE`, numeric vector is returned.
 #'
-#' @author C. Beleites
+#'
+#' @return
+#' - If `index = FALSE`, function returns a `hyperSpec` object, `data.frame` or
+#'   `matrix` (object of the same class as `x`) with `size` rows.
+#' - If `index = TRUE`, function returns a vector with row indices of size
+#'   `size` suitable for subsetting rows of `x`.
+#'
+#' @seealso [base::sample()]
+#'
+#'
+#' @author C. Beleites, V. Gegzna
 #'
 #' @keywords methods distribution
 #' @concept stats
 #'
 #' @export
 #'
-#' @seealso [base::sample()]
 #' @examples
+#' set.seed(2021)
 #'
 #' sample(flu, 3)
+#'
+#' sample(flu, 3, index = TRUE)
+#'
+#' sample(flu, 3, replace = TRUE, index = TRUE)
+#'
+#' sample(flu, 8, replace = TRUE, index = TRUE)
+#'
 #'
 #' plot(flu, col = "darkgray")
 #' plot(sample(flu, 3), col = "red", add = TRUE)
@@ -52,15 +78,16 @@
 #'   col = "#0000FF80", add = TRUE,
 #'   lines.args = list(lwd = 2)
 #' )
-setMethod("sample", signature = signature(x = "hyperSpec"), .sample)
+setMethod("sample", signature = signature(x = "hyperSpec"), .sample_h)
 
 # Unit tests -----------------------------------------------------------------
 
-hySpc.testthat::test(.sample) <- function() {
+hySpc.testthat::test(.sample_h) <- function() {
   context(".sample")
 
   test_that("defaults", {
     tmp <- sample(flu)
+    expect_s4_class(tmp, "hyperSpec")
     expect_equal(tmp[order(tmp$c)], flu)
 
     set.seed(101)
@@ -68,16 +95,47 @@ hySpc.testthat::test(.sample) <- function() {
   })
 
   test_that("size", {
-    expect_length(isample(flu, size = 3), 3L)
+    expect_length(sample(flu, size = 3), 3L)
   })
 
   test_that("prob", {
-    expect_equal(isample(flu, size = 1, prob = c(1, rep(0, 5))), 1L)
+    expect_equal(sample(flu, size = 1, prob = c(1, rep(0, 5))), flu[1L])
   })
 
   test_that("replace", {
     expect_equal(
-      isample(flu, size = 3, replace = TRUE, prob = c(1, rep(0, 5))),
+      sample(flu, size = 3, replace = TRUE, prob = c(1, rep(0, 5))),
+      flu[rep(1L, 3)]
+    )
+  })
+
+
+  test_that("defaults, index = TRUE", {
+    tmp <- sort(sample(flu, index = TRUE))
+    expect_equal(tmp, 1:nrow(flu))
+    expect_type(tmp, "integer")
+    expect_equal(class(tmp), "integer")
+
+    set.seed(101)
+    expect_equal(sample(flu, index = TRUE), c(1L, 6L, 2L, 3L, 5L, 4L))
+  })
+
+  test_that("size, index = TRUE", {
+    expect_length(sample(flu, size = 3, index = TRUE), 3L)
+  })
+
+  test_that("prob, index = TRUE", {
+    expect_equal(
+      sample(flu, size = 1, prob = c(1, rep(0, 5)), index = TRUE),
+      1L
+    )
+  })
+
+  test_that("replace, index = TRUE", {
+    probs <- c(1, rep(0, 5))
+
+    expect_equal(
+      sample(flu, size = 3, replace = TRUE, prob = probs, index = TRUE),
       rep(1L, 3)
     )
   })
@@ -87,16 +145,23 @@ hySpc.testthat::test(.sample) <- function() {
 # Function -------------------------------------------------------------------
 
 .sample.data.frame <- function(x, size, replace = FALSE, prob = NULL,
-                               drop = FALSE) {
+                               index = FALSE, drop = FALSE) {
   if (missing(size)) size <- nrow(x)
   rows_i <- sample.int(nrow(x), size = size, replace = replace, prob = prob)
-  x[rows_i, , drop = drop]
+
+  if (isTRUE(index)) {
+    rows_i
+  } else {
+    x[rows_i, , drop = drop]
+  }
 }
 
 #' @rdname sample
+#' @aliases sample.data.frame
+#'          sample,data.frame-method
 #'
-#' @param drop See [base::drop()]: by default, do not drop dimensions of the
-#'        result.
+#' @param drop (logical): See [base::drop()], by default, do not drop dimensions
+#'        of the result. Applicable only if `index = FALSE`.
 #'
 #' @concept stats
 #'
@@ -105,6 +170,8 @@ hySpc.testthat::test(.sample) <- function() {
 #' @examples
 #'
 #' sample(cars, 2)
+#'
+#' sample(cars, 2, index = TRUE)
 setMethod("sample", signature = signature(x = "data.frame"), .sample.data.frame)
 
 
@@ -115,6 +182,7 @@ hySpc.testthat::test(.sample.data.frame) <- function() {
   test_that("data.frame", {
     set.seed(101)
     tmp <- sample(iris)
+    expect_s3_class(tmp, "data.frame")
     expect_equal(rownames(tmp), c(
       "73", "57", "95", "148", "61", "59", "99", "128", "131", "32",
       "9", "96", "144", "98", "60", "147", "145", "14", "97", "45",
@@ -135,25 +203,44 @@ hySpc.testthat::test(.sample.data.frame) <- function() {
     expect_equal(dim(tmp), dim(iris))
     expect_equal(tmp, iris[as.numeric(rownames(tmp)), ])
   })
+
+  test_that("data.frame, index = TRUE", {
+    set.seed(101)
+    tmp <- sample(iris, index = TRUE)
+    expect_type(tmp, "integer")
+    expect_equal(class(tmp), "integer")
+    expect_length(tmp, nrow(iris))
+    expect_equal(tmp[1:3], c(73, 57, 95))
+  })
 }
 
 
 # Function -------------------------------------------------------------------
 
-.sample.matrix <- function(x, size, replace = FALSE, prob = NULL, drop = FALSE) {
+.sample.matrix <- function(x, size, replace = FALSE, prob = NULL,
+                           index = FALSE, drop = FALSE) {
   if (missing(size)) size <- nrow(x)
   rows_i <- sample.int(nrow(x), size = size, replace = replace, prob = prob)
-  x[rows_i, , drop = drop]
+
+  if (isTRUE(index)) {
+    rows_i
+  } else {
+    x[rows_i, , drop = drop]
+  }
 }
 
 #' @rdname sample
-#' @export
+#' @aliases sample.matrix
+#'          sample,matrix-method
 #'
 #' @concept stats
 #'
+#' @export
 #' @examples
 #'
 #' sample(matrix(1:24, 6), 2)
+#'
+#' sample(matrix(1:24, 6), 2, index = TRUE)
 setMethod("sample", signature = signature(x = "matrix"), .sample.matrix)
 
 
@@ -164,66 +251,19 @@ hySpc.testthat::test(.sample.matrix) <- function() {
   test_that("matrix", {
     set.seed(101)
     tmp <- sample(flu[[]])
+    expect_equal(class(tmp)[1], "matrix")
+
     expect_equal(dim(tmp), dim(flu[[]]))
     expect_equal(tmp[c(1L, 3L, 4L, 6L, 5L, 2L), ], flu[[]])
   })
-}
 
-
-# Function -------------------------------------------------------------------
-
-#' @rdname sample
-#'
-#' @description
-#' [isample()] returns an vector of indices, `sample()` returns the
-#' corresponding `hyperSpec` object.
-#'
-#' @return
-#' [isample()] returns vector with indices suitable for row-indexing `x`.
-#'
-#' @export
-#'
-#' @concept stats
-#'
-#' @examples
-#'
-#' isample(flu, 3)
-#'
-#' isample(flu, 3, replace = TRUE)
-#'
-#' isample(flu, 8, replace = TRUE)
-isample <- function(x, size = nrow(x), replace = FALSE, prob = NULL) {
-  assert_hyperSpec(x)
-  validObject(x)
-
-  sample.int(nrow(x), size = size, replace = replace, prob = prob)
-}
-
-
-# Unit tests -----------------------------------------------------------------
-
-hySpc.testthat::test(isample) <- function() {
-  context("isample")
-
-  test_that("defaults", {
-    expect_equal(sort(isample(flu)), 1:nrow(flu))
-
+  test_that("matrix, index = TRUE", {
     set.seed(101)
-    expect_equal(isample(flu), c(1L, 6L, 2L, 3L, 5L, 4L))
-  })
+    tmp <- sample(flu[[]], index = TRUE)
+    expect_equal(class(tmp), "integer")
+    expect_type(tmp, "integer")
 
-  test_that("size", {
-    expect_equal(nrow(sample(flu, size = 3)), 3L)
-  })
-
-  test_that("prob", {
-    expect_equal(sample(flu, size = 1, prob = c(1, rep(0, 5))), flu[1L])
-  })
-
-  test_that("replace", {
-    expect_equal(
-      sample(flu, size = 3, replace = TRUE, prob = c(1, rep(0, 5))),
-      flu[rep(1L, 3)]
-    )
+    expect_equal(length(tmp), nrow(flu[[]]))
+    expect_equal(tmp,  c(1, 6, 2, 3, 5, 4))
   })
 }